+++ src/option.ts
@@ -3,6 +3,7 @@ export type BuilderOption = {
   addAscAfterOrderByItem?: 'always'|'never'|'when mentioned'
   addNullsLastAfterOrderByItem?: 'always'|'never'|'when mentioned'
   addAsBeforeColumnAlias?: 'always'|'never'
+  addPublicSchemaName?: 'always'|'never'|'when other schema mentioned'
 }
 
 const defaultOption: BuilderOption = {
@@ -10,6 +11,7 @@ const defaultOption: BuilderOption = {
   addAscAfterOrderByItem: 'when mentioned',
   addNullsLastAfterOrderByItem: 'when mentioned',
   addAsBeforeColumnAlias: 'always',
+  addPublicSchemaName: 'never'
 }
 Object.freeze(defaultOption)
 
@@ -19,5 +21,6 @@ export function fillUndefinedOptionsWithDefault(option: BuilderOption): BuilderO
   result.addAscAfterOrderByItem = option.addAscAfterOrderByItem ?? defaultOption.addAscAfterOrderByItem
   result.addNullsLastAfterOrderByItem = option.addNullsLastAfterOrderByItem ?? defaultOption.addNullsLastAfterOrderByItem
   result.addAsBeforeColumnAlias = option.addAsBeforeColumnAlias ?? defaultOption.addAsBeforeColumnAlias
+  result.addPublicSchemaName = option.addPublicSchemaName ?? defaultOption.addPublicSchemaName
   return result
 }