+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Rename models to types and move it inside models folder
 - Rename file select.ts to SelectItemInfo.ts
 - Add IStatementGiver Interface
+- Add option addPublicSchemaName to Builder Option
+- Add Builder data to Table getStmt
 
 ### Added
 - Having step can contain aggregate function condition